extract downloadwatchers
